Add support for setting additional annotations on the CRDs
This is especially useful since CRDs are now installed through templates.
Because of this change that was introduced in 0.57.0, now running
`helm uninstall` of the operator chart will delete the CRDs as well,
which is a destructive operation.

The workaround is to leverage the special helm.sh/resource-policy=keep
annotation, see https://helm.sh/docs/howto/charts_tips_and_tricks/#tell-helm-not-to-uninstall-a-resource

One currently open issue though, is that the annotation needs to be set
by Helm itself, see https://github.com/helm/helm/issues/10890
So running kubectl annotate on the CRDs as a post-install step doesn't work.